Renamed package 'dao' to 'repositories' to match Spring naming conventions. Made tests for guestService
