+++ src/main/java/no/hvl/dat250/h2020/group5/controllers/GuestController.java
@@ -18,7 +18,7 @@ public class GuestController {
 
     @GetMapping()
     public List<Guest> getAllGuests(){
-        return guestService.getAllGuest();
+        return guestService.getAllGuests();
     }
 
     @PostMapping()
