+++ src/test/java/no/hvl/dat250/h2020/group5/repostories/VoteRepositoryTest.java
@@ -1,9 +1,9 @@
 package no.hvl.dat250.h2020.group5.repostories;
 
 import no.hvl.dat250.h2020.group5.Main;
-import no.hvl.dat250.h2020.group5.dao.PollRepository;
-import no.hvl.dat250.h2020.group5.dao.VoteRepository;
-import no.hvl.dat250.h2020.group5.dao.VoterRepository;
+import no.hvl.dat250.h2020.group5.repositories.PollRepository;
+import no.hvl.dat250.h2020.group5.repositories.VoteRepository;
+import no.hvl.dat250.h2020.group5.repositories.VoterRepository;
 import no.hvl.dat250.h2020.group5.entities.Guest;
 import no.hvl.dat250.h2020.group5.entities.Poll;
 import no.hvl.dat250.h2020.group5.entities.Vote;
